+++ CHANGES.md
@@ -11,8 +11,10 @@ Beta 5 (in progress)
 * added show() and hide()
 * added define() (or, rather, documented it)
 * added stop() for all promises, to propagate animation cancellation
+* added $$show property for set() and get()
 * removed delay() and defer(), as they just duplicated setTimeout()
 * clone() will now use the browser's cloneNode(). It will only remove the element from the top-level element, but not from any children.
+* $$slide and $$fade's behaviour has slightly changed (will not modify $display anymore), please see the API docs.
 * formatValue() supports any character as group separator
 * updated Minified's promise implementation to the latest Promises/A+ spec (full package only)
 * a promise's state can now be obtained by invoking it as a function without arguments (see promise(), full package only). promise.state has been removed.